Model parameters

adap is the only variable exported from MKtest module. It is a Mutable structure contaning the variables required to solve the analytical approach. Any value can be easly changed. Remember adap should be change before the execution, in other case, $\alpha_{(x)}$ will be solve with the default values. To change all the values at once, you can use MKtest.parameters in order to set specific models in a new variable.

MKtest.parameters — Type

Mutable structure containing the variables required to solve the analytical approach. All the functions are solve using the internal values of the structure. You should declare a mutable structure to the perform the analytical estimations.

Parameters

gam_neg::Int64: Selection coefficient for deleterious alleles

gL::Int64: Selection coefficient for weakly benefical alleles

gH::Int64: Selection coefficient for strongly benefical alleles

al_low::Float64: Proportion of α due to weak selection

al_tot::Float64: α

θ_noncoding::Float64: Mutation rate defining BGS strength

θ_coding::Float64: Mutation rate on coding region

al::Float64: DFE shape parameter

be::Float64: DFE scale parameter

B::Float64: BGS strength

B_bins::Array{Float64,1}: BGS values to simulate

ppos_l::Float64: Fixation probabily of weakly beneficial alleles

ppos_h::Float64: Fixation probabily of strongly beneficial alleles

N::Int64: Population size

n::Int64: Sample size

Lf::Int64: Flanking region length

ρ::Float64: Recombination rate

TE::Float64

MKtest.Br — Function

Br(Lmax,theta)

Expected reduction in nucleotide diversity. Explored at Charlesworth B., 1994:

\[\frac{\pi}{\pi_{0}} = e^{\frac{4\muL}{2rL+t}}\]

Arguments

param::parameters

theta::Float64

Returns

Float64: expected reduction in diversity given a non-coding length, mutation rate and defined recombination.

MKtest.set_θ! — Function

set_θ!(param)

Find the optimum mutation given the expected reduction in nucleotide diversity (B value) in a locus.

Returns

adap.θ_noncoding::Float64: changes adap.θ_noncoding value.

MKtest.set_ppos! — Function

set_ppos!(param)

Find the probabilty of positive selected alleles given the model. It solves a equation system taking into account fixations probabilities of weakly and strong beneficial alleles.

Returns

Tuple{Float64,Float64}: weakly and strong beneficial alleles probabilites.

MKtest.binom_op! — Function

binom_op(param)

Binomial convolution to sample the allele frequencies probabilites depending on background selection values and sample size.

Arguments

param::parameters

binom::binomialDict

Returns

Array{Float64,2}: convoluted SFS for each B value defined in the model (param.B_bins). The estimations are saved at convolutedBn.bn.

MKtest.Φ — Function

Φ(param,gamma,ppos)

Reduction in fixation probabilty due to background selection and linkage. The formulas used have been subjected to several theoretical works (Charlesworth B., 1994, Hudson et al., 1995, Nordborg et al. 1995, Barton NH., 1995).

The fixation probabilty of selected alleles are reduce by a factor $\phi$:

\[\phi(t,s) = e^{[\frac{-2\mu}{t(1+\frac{rL}{t}+\frac{2s}{t})}]}\]

Multiplying across all deleterious linkes sites, we find:

\[\Phi = \prod_{1}^{L} = \phi(t,s) = e^{\frac{-2t\mu(\psi[1,\frac{r+2s+L}{r}] - \psi[1,\frac{r(L+1)+2s+t}{r}])}{r^2}}\]

\[\phi(t,s) = e^{\frac{-2t\mu(\psi[1,\frac{r+2s+L}{r}] - \psi[1,\frac{r(L+1)+2s+t}{r}])}{r^2}}\]

Arguments

gamma::Int64: selection coefficient.

Returns

Float64: expected rate of positive fixations under background selection.

MKtest.analytical_alpha — Function

analyticalAlpha(param, binom)

Analytical α(x) estimation. Solve α(x) generally. We used the expected rates of divergence and polymorphism to approach the asympotic value accouting for background selection, weakly and strong positive selection. α(x) can be estimated taking into account the role of positive selected alleles or not. In this way we explore the role of linkage to deleterious alleles in the coding region.

\[\mathbb{E}[\alpha_{x}] = 1 - \left(\frac{\mathbb{E}[D_{s}]}{\mathbb{E}[D_{N}]}\frac{\mathbb{E}[P_{N}]}{\mathbb{E}[P_{S}]}\right)\]

Arguments

param::parameters

Returns

Array{Float64,1} α(x).

MKtest estimation

Fixations

MKtest.fix_neut — Function

fix_neut()

Expected neutral fixations rate reduce by B value.

\[\mathbb{E}[D_{s}] = (1 - p_{-} - p_{+}) B \frac{1}{2N}\]

Returns

Float64: expected rate of neutral fixations.

Missing docstring.

Missing docstring for MKtest.fix_neg_b. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MKtest.p_fix. Check Documenter's build log for details.

MKtest.fix_pos_sim — Function

fix_pos_sim(gamma,ppos)

Expected positive fixations rate reduced due to the impact of background selection and linkage. The probabilty of fixation of positively selected alleles is reduced by a factor Φ across all deleterious linked sites Analytical.Φ.

\[\mathbb{E}[D_{n+}] = \Phi \cdot \mathbb{E}[D_{n+}]\]

Arguments

ppos::Float64: positive selected alleles probabilty.

Returns

Float64: expected rate of positive fixations under background selection

Polymorphism

MKtest.sfs_neut — Function

sfs_neut()

Expected rate of neutral allele frequency reduce by backgrou nd selection. The spectrum depends on the number of individual []

\[\mathbb{E}[Ps_{(x)}] = \sum{x^{*}=x}{x^{*}=1}f_{B}(x)\]

Input

param::parameters: mutable structure containing the model

SparseMatrixCSC{Float64, Int64}: convoluted SFS given the B value.

Return:

Array{Float64}: expected rate of neutral alleles frequencies.

MKtest.sfs_pos — Function

sfs_pos(s,p)

Expected rate of positive selected allele frequency reduce by background selection. The spectrum depends on the number of individuals.

Arguments

param::parameters: mutable structure containing the model.

s::Int64: selection strength.

p::Float64: positive selected alleles probabilty.

binom::SparseMatrixCSC{Float64, Int64}: convoluted SFS given the B value.

Return:

Array{Float64}: expected positive selected alleles frequencies.

MKtest.sfs_neg — Function

sfs_neg(param,p)

Expected rate of positive selected allele frequency reduce by background selection. Spectrum drawn on a gamma DFE. It depends on the number of individuals.

Arguments

param::parameters: mutable structure containing the model

p::Float64: positive selected alleles probabilty.

binom::SparseMatrixCSC{Float64, Int64}: convoluted SFS given the B value.

Return:

Array{Float64}: expected negative selected alleles frequencies.

MKtest.cumulative_sfs — Function

cumulative_sfs(sfs_tmp)

Changing SFS considering all values above a frequency x. The original asymptotic-MK approach takes Pn(x) and Ps(x) as the number of polymorphic sites at frequency x rather than above x, but this approach scales poorly as sample size increases. We define the polymorphic spectrum as stated above since these quantities trivially have the same asymptote but are less affected by changing sample size.

Arguments

sfs_tmp::Vector{Float64}: SFS array

freqs::Bool: true/false wheter the input SFS array containing a first columns with the frequencies.

Output

sfs_tmp::Vector{Float64}: SFS array

Rates

MKtest.rates — Function

rates(param::parameters,iterations::Int64,divergence::Array,sfs::Array)

Function to solve randomly N scenarios. The function will create N models, defined by Analytical.parameters(), to estimate analytically fixation and polymorphic rates for each model. The rates will be used to compute summary statistics required at ABC. The function output a HDF5 file containing the solved models, the selected DAC and the analytical rates.

If ρ and/or theta are set to nothing, the function will input random values given the range 0.0005:0.0005:0.01. Otherwise you can fix the values.

If gL is set to nothing, the function will not account the role of the weakly selected alleles in the estimation.

Arguments

param::parameters: mutable structure containing the model

binom::binomialDict : structure containing the binomial convolution

gH::Array{Int64,1} : Range of strong selection coefficients

gL::Union{Array{Int64,1},Nothing}: Range of weak selection coefficients

gam_neg::Array{Int64,1} : Range of deleterious selection coefficients

theta::Union{Float64,Nothing} : Population-scaled mutation rate on coding region

ρ::Union{Float64,Nothing} : Population-scaled recombination rate

shape::Float64=0.184 : DFE shape parameter

iterations::Int64 : Number of solutions

output::String : File to output HDF5 file

Returns

Array: summary statistics

Output: HDF5 file containing models solved and rates.

MKtest.iter_rates — Function

iter_rates(param::parameters,afac::Float64,bfac::Float64,al_tot::Float64,al_low::Float64,divergence::Array,sfs::Array)

Estimating rates given a model for all B range.

Arguments

param::parameters

binom::Dict

al_tot::Float64

al_low::Float64

gH::Int64

gL::Int64

gam_neg::Int64

afac::Float64

ρ::Float64

θ::Float64

Output

Array{Float64,2}

MKtest.getting_rates — Function

getting_rates(gammaL,gammaH,ppos_l,ppos_h,observedData,nopos)

Estimating analytical rates of fixation and polymorphism to approach α value accouting for background selection, weakly and strong positive selection. Output values will be used to sample from a Poisson distribution the total counts of polymorphism and divergence using observed data.

Arguments

param::parameters

binom::SparseMatrixCSC{Float64,Int64}

Returns

Array{Float64,2} containing solved model, fixation and polymorphic rates

Summary statistics

MKtest.poisson_fixation — Function

poisson_fixation(observed_values,λds, λdn)

Divergence sampling from Poisson distribution. The expected neutral and selected fixations are subset through their relative expected rates (fix_neut, fix_neg_b, fix_pos_sim). Empirical values are used are used to simulate the locus L along a branch of time T from which the expected Ds and Dn raw count estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily given the success rate $\lambda$ in the distribution.

\[\mathbb{E}[D_N] = X \in Poisson\left(\lambda = D \times \left[\frac{\mathbb{E}[D_+] + \mathbb{E}[D_-]}{\mathbb{E}[D_+] + \mathbb{E}[D_-] + \mathbb{E}[D_0]}\right]\right)\]

\[\mathbb{E}[D_S] = X \in Poisson\left(\lambda = D \times \left[\frac{\mathbb{E}[D_0]}{\mathbb{E}[D_+] + \mathbb{E}[D_-] + \mathbb{E}[D_0]}\right]\right)\]

Arguments

observed_values::Array: Array containing the total observed divergence.

λds::Float64: expected neutral fixations rate.

λdn::Float64: expected selected fixations rate.

Returns

Array{Int64,1} containing the expected count of neutral and selected fixations.

MKtest.poisson_polymorphism — Function

poisson_polymorphism(observed_values,λps,λpn)

Polymorphism sampling from Poisson distributions. The total expected neutral and selected polimorphism are subset through the relative expected rates at the frequency spectrum (fix_neut, sfs_neut,). Empirical sfs are used to simulate the locus L along a branch of time T from which the expected Ps and Pn raw count are estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily from the whole sfs given each frequency success rate $\lambda$ in the distribution.

The success rate managing the Poisson distribution by the observed count each frequency. We considered both sampling variance and process variance is affecting the number of variable alleles we sample from SFS. This variance arises from the random mutation-fixation process along the branch. To incorporate this variance we do one sample per frequency-bin and use the total sampled variation and the SFS for the summary statistics.

\[\mathbb{E}[P_N] = \sum_{x=0}^{x=1} X \in Poisson\left(\lambda = SFS_{(x)} \times \left[\frac{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}]}{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}] + \mathbb{E}[P_{0(x)}]}\right]\right)\]

\[\mathbb{E}[P_S] = \sum_{x=0}^{x=1} X \in Poisson\left(\lambda = SFS_{(x)} \times \left[\frac{\mathbb{E}[P_{0(x)}]}{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}] + \mathbb{E}[P_{0(x)}]}\right]\right)\]

Arguments

observed_values::Array{Int64,1}: Array containing the total observed divergence.

λps::Array{Float64,1}: expected neutral site frequency spectrum rate.

λpn::Array{Float64,1}: expected selected site frequency spectrum rate.

Returns

Array{Int64,2} containing the expected total count of neutral and selected polymorphism.

MKtest.sampled_from_rates — Function

sampled_alpha(observed_values,λds, λdn)

Ouput the expected values from the Poisson sampling process. Please check poisson_fixation and poisson_polymorphism to understand the samplingn process. α(x) is estimated through the expected values of Dn, Ds, Pn and Ps.

Arguments

param::parameters: Array containing the total observed divergence.

d::Array: observed divergence.

afs::Array: observed polymorphism.

λdiv::Array{Float64,2}: expected fixations rate.

λdiv::Array{Float64,2}: expected site frequency spectrum rates.

Returns

αsummaries,expdn,expds,exppn,exp_ps,ssAlpha

Array{Int64,2} containing α(x) values.

Array{Int64,1} expected non-synonymous divergence.

Array{Int64,1} expected synonymous divergence.

Array{Int64,1} expected non-synonymous polymorphism.

Array{Int64,1} expected synonymous polymorphism.

Array{Int64,1} expected synonymous polymorphism.

Array{Int64,1} expected synonymous polymorphism.

Array{Int64,2} containing α(x) binned values.

sampled_from_rates(gammaL,gammaH,ppos_l,ppos_h,observedData,nopos)

MKtest.summary_statistics — Function

summary_statistics(param::parameters,rates::JLD2.JLDFile,analysis_folder::String,summstat_size::Int64,replicas::Int64,bootstrap::Bool)

Estimate summary statistics using observed data and analytical rates. analysis_folder will check for the SFS and divergence file and will be used to output summary statistics

Arguments

param::parameters : Mutable structure containing the models

h5_file::String : HDF5 containing solved models, fixation and polymorphic rates

sfs::Vector{Matrix{Float64}}: SFS matrix.

divergence::Vector{Matrix{Float64}} : divergence matrix.

summstat_size::Int64 : Number of summary statistics

replicas::Int64 : Number of bootstrap replicas

bootstrap::Bool : Boolean to perform or not bootstrapping

Output

Summary statistics to ABC inference

Inference tools

MKtest.parse_sfs — Function

parse_sfs(;sample_size::Int64,data::String)

Function to parse polymorphism and divergence by subset of genes. The input data is based on supplementary material described at Uricchio et al. 2019. Please be sure the file is tabulated.

GeneId	Pn	DAF seppareted by commas	Ps	DAF separated by commas	Dn	Ds

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

Arguments

sample_size::Int64: data sample size.

data: File containing polymorphic and divergence data

gene_list: File containing gene IDs to subset. You can perform multiple subset input a file where eachrow contain the gene IDs to subset.

sfs_columns::Array{Int64,1}: non-synonymous and synonymous daf columns. Please introduce first the non-synonymous number.

div_columns::Array{Int64,1}: non-synonymous and synonymous divergence columns. Please introduce first the non-synonymous number.

bins::Array{Int64,1}: bin to collapse the SFS from sample_size to new bin size

Returns

Array{Float64,1}: α values

Array{Float64,2}: Site Frequency Spectrum

Array{Float64,1}: Synonymous and non-synonymous divergence counts

MKtest.ABCreg — Function

ABCreg(analysis_folder, replicas, S, tol, abcreg)

Performing ABC inference using ABCreg. Please, be sure your analysis_folder contain the files alphas.txt and summaries.txt produced by Analytical.summaryStatsFromRates()

Arguments

analysis_folder::String : Folder containing the observed data and summary estatistics. It will be used to output the posterior distributions

P::Int64 : Number of parameters to perform the inference.

S::Int64 : Number of summary stastitics to perform the inference.

tol::Float64 : Tolerance value. It define the number of accepted values at ABC inference

abcreg::String : Path to ABCreg binary

Output

Files containing posterior distributions from ABCreg

MKtest.source_plot_map_r — Function

Function to download and source plotMap function. We do not include at part of the module to avoid external dependecies. Once the function is execute properly you will have a function called *plotMap which used R to 
	estimate and plot the Maximum A Posterior following ABCreg example. It uses locfit and ggplot2 libraries.

#Arguments

script_path::String: Path and name to save MAP scritp.

MKtest

MKtest.aMK — Function

aMK(x,columns)

Function to estimate the asymptotic value of α(x).

Arguments

alpha_values::Array{Float64,1}: α(x) array.

Returns

Array{Float64,2}: Array of array containing asymptotic values, lower confidence interval, higher confidence interval

MKtest.imputedMK — Function

imputedMK(sfs,divergence,m,cutoff)

Function to estimate the imputedMK

Arguments

sfs::Matrix{Float64}: SFS array

divergence::Array: divegence count array

m::Union{Nothing,Array}: non-synonymous and synonymous sites

cutoff{Float64}: frequency cutoff to perform imputedMK

Returns

`Dict: Dictionary containing results

MKtest.fwwMK — Function

fwwMK(sfs,divergence,m,cutoff)

Function to estimate the Fay, Wycoff and Wu's MK extension.

Arguments

sfs::Matrix{Float64}: SFS array

divergence::Array: divegence count array

m::Union{Nothing,Array}: non-synonymous and synonymous sites

cutoff{Float64}: frequency cutoff to perform fwwMK

Returns

`Dict: Dictionary containing results

MKtest.standardMK — Function

standardMK(x,columns)

Function to estimate the original α value

Arguments

sfs::Matrix{Float64}: SFS array

divergence::Array: divegence count array

m::Union{Nothing,Array}: non-synonymous and synonymous sites# Returns

Output

`Dict: Dictionary containing results